[gtk] -Change: xy event delivers coords in long (strictly speaking: an int can be 16 bit)
git-svn-id: svn://repo.hu/librnd/trunk@33292 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/librnd/plugins/lib_gtk_common/rnd_gtk.h
@@ -103,7 +103,7 @@ struct rnd_gtk_mouse_s {
 };
 
 typedef struct { /* drawing area resize event binding; compat.h will implement this */
-	gboolean (*cb)(GtkWidget *widget, int x, int y, void *user_data);
+	gboolean (*cb)(GtkWidget *widget, long x, long y, void *user_data);
 	void *user_data;
 } gtkc_event_xy_t;
